// src/@noctua.core/models/relations.ts
export enum Relations {
  ENABLED_BY = "RO:0002333",
  HAPPENS_DURING = 'RO:0002092',
  LOCATED_IN = 'RO:0001025',
  OCCURS_IN = 'BFO:0000066',
  PART_OF = 'BFO:0000050',
  CAUSALLY_UPSTREAM_OF = 'RO:0002411',
  CAUSALLY_UPSTREAM_OF_OR_WITHIN = 'RO:0002418',
  CAUSALLY_UPSTREAM_OF_POSITIVE_EFFECT = 'RO:0002304',
  CAUSALLY_UPSTREAM_OF_NEGATIVE_EFFECT = 'RO:0002305',
  CAUSALLY_UPSTREAM_OF_OR_WITHIN_POSITIVE_EFFECT = 'RO:0004047',
  CAUSALLY_UPSTREAM_OF_OR_WITHIN_NEGATIVE_EFFECT = 'RO:0004046',
  CONSTITUTIVELY_UPSTREAM_OF = 'RO:0012009',
  DIRECTLY_NEGATIVELY_REGULATES = 'RO:0002630',
  DIRECTLY_POSITIVELY_REGULATES = 'RO:0002629',
  HAS_INPUT = 'RO:0002233',
  HAS_OUTPUT = 'RO:0002234',
  INDIRECTLY_NEGATIVELY_REGULATES = 'RO:0002409',
  INDIRECTLY_POSITIVELY_REGULATES = 'RO:0002407',
  IS_SMALL_MOLECULE_INHIBITOR_OF = 'RO:0012006',
  IS_SMALL_MOLECULE_ACTIVATOR_OF = 'RO:0012005',
  NEGATIVELY_REGULATES = 'RO:0002212',
  POSITIVELY_REGULATES = 'RO:0002213',
  PROVIDES_INPUT_FOR = 'RO:0002413',
  REMOVES_INPUT_FOR = 'RO:0012010',
}